Condensed utility files into perception combined utility files
