+++ GEMstack/onboard/perception/combined_detection.py
@@ -1,8 +1,8 @@
 from ...state import AllState, VehicleState, ObjectPose, ObjectFrameEnum, AgentState, AgentEnum, AgentActivityEnum
 from ..interface.gem import GEMInterface
 from ..component import Component
-from .pedestrian_utils import pose_to_matrix
-from .pedestrian_utils_gem import *
+from .peerception_utils import pose_to_matrix
+from .perception_utils_gem import *
 from typing import Dict, List, Optional, Tuple
 import rospy
 from message_filters import Subscriber, ApproximateTimeSynchronizer
