+++ GEMstack/onboard/perception/perception_utils.py
@@ -1,4 +1,3 @@
-from ...state import ObjectPose, AgentState
 from typing import Dict
 import open3d as o3d
 import numpy as np
@@ -7,6 +6,7 @@ from scipy.spatial.transform import Rotation as R
 from sensor_msgs.msg import PointCloud2
 import sensor_msgs.point_cloud2 as pc2
 import ros_numpy
+import math
 
 
 # ----- Helper Functions -----
@@ -22,37 +22,6 @@ def filter_points_within_threshold(points, threshold=15.0):
     distances = np.linalg.norm(points, axis=1)
     mask = distances <= threshold
     return points[mask]
-
-def match_existing_cone(
-        new_center: np.ndarray,
-        new_dims: tuple,
-        existing_agents: Dict[str, AgentState],
-        distance_threshold: float = 1.0
-) -> str:
-    """
-    Find the closest existing Cone agent within a specified distance threshold.
-    """
-    best_agent_id = None
-    best_dist = float('inf')
-    for agent_id, agent_state in existing_agents.items():
-        old_center = np.array([agent_state.pose.x, agent_state.pose.y, agent_state.pose.z])
-        dist = np.linalg.norm(new_center - old_center)
-        if dist < distance_threshold and dist < best_dist:
-            best_dist = dist
-            best_agent_id = agent_id
-    return best_agent_id
-
-
-def compute_velocity(old_pose: ObjectPose, new_pose: ObjectPose, dt: float) -> tuple:
-    """
-    Compute the (vx, vy, vz) velocity based on change in pose over time.
-    """
-    if dt <= 0:
-        return (0, 0, 0)
-    vx = (new_pose.x - old_pose.x) / dt
-    vy = (new_pose.y - old_pose.y) / dt
-    vz = (new_pose.z - old_pose.z) / dt
-    return (vx, vy, vz)
 
 
 def extract_roi_box(lidar_pc, center, half_extents):
@@ -224,3 +193,27 @@ def project_points(pts_cam, K, original_lidar_points):
     v = (K[1, 1] * (Yc / Zc) + K[1, 2]).astype(np.int32)
     proj = np.column_stack((u, v, lidar_valid))
     return proj
+
+def pose_to_matrix(pose):
+    """
+    Compose a 4x4 transformation matrix from a pose state.
+    Assumes pose has attributes: x, y, z, yaw, pitch, roll,
+    where the angles are given in degrees.
+    """
+    # Use default values if any are None (e.g. if the car is not moving)
+    x = pose.x if pose.x is not None else 0.0
+    y = pose.y if pose.y is not None else 0.0
+    z = pose.z if pose.z is not None else 0.0
+    if pose.yaw is not None and pose.pitch is not None and pose.roll is not None:
+        yaw = math.radians(pose.yaw)
+        pitch = math.radians(pose.pitch)
+        roll = math.radians(pose.roll)
+    else:
+        yaw = 0.0
+        pitch = 0.0
+        roll = 0.0
+    R_mat = R.from_euler('zyx', [yaw, pitch, roll]).as_matrix()
+    T = np.eye(4)
+    T[:3, :3] = R_mat
+    T[:3, 3] = np.array([x, y, z])
+    return T
