+++ GEMstack/onboard/perception/yolo_node.py
@@ -1,4 +1,4 @@
-from pedestrian_utils import *
+from perception_utils import *
 from combined_detection_utils import add_bounding_box
 from ultralytics import YOLO
 import cv2
